load-gen/requirements.txt
psycopg2-binary==2.9.5
numpy==1.23.3
uvicorn==0.20.0
fastapi==0.87.0